requirements.txt
numpy==1.16.6
pandas==1.2.4
scikit-learn==0.24.2
scikit-survival==0.15.0
tensorflow-gpu==1.14.0
